requirements.txt
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cymem==2.0.5
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.0.0/de_core_news_sm-3.0.0-py3-none-any.whl
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
jellyfish==0.6.0
Jinja2==3.0.1
joblib==1.0.1
langdetect==1.0.9
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pathy==0.5.2
pdf2image==1.14.0
pdfminer.six==20201018
Pillow==8.2.0
preshed==3.0.5
pycparser==2.20
pydantic==1.7.4
pyparsing==2.4.7
pytesseract==0.3.7
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
smart-open==3.0.0
sortedcontainers==2.4.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
thinc==8.0.3
timeloop==1.0.2
tqdm==4.60.0
typer==0.3.2
urllib3==1.26.5
waitress==2.0.0
wasabi==0.8.2
Werkzeug==2.0.1
